# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "FastAPI",
        "version": "0.1.0"
    },
    "paths": {
...
        "/api/main/mixtape/": {
            "post": {
                "tags": [
                    "mixtape"
                ],
                "summary": "Create Mixtape",
                "operationId": "create_mixtape_api_main_mixtape__post",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MixtapeRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": true,
                                    "title": "Response Create Mixtape Api Main Mixtape  Post"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "tags": [
                    "mixtape"
                ],
                "summary": "List My Mixtapes",
                "operationId": "list_my_mixtapes_api_main_mixtape__get",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Search mixtape titles (partial match)",
                            "title": "Q"
                        },
                        "description": "Search mixtape titles (partial match)"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1,
                            "description": "Max results to return",
                            "default": 20,
                            "title": "Limit"
                        },
                        "description": "Max results to return"
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 0,
                            "description": "Results offset for pagination",
                            "default": 0,
                            "title": "Offset"
                        },
                        "description": "Results offset for pagination"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "additionalProperties": true
                                    },
                                    "title": "Response List My Mixtapes Api Main Mixtape  Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/main/mixtape/{public_id}/claim": {
            "post": {
                "tags": [
                    "mixtape"
                ],
                "summary": "Claim Mixtape",
                "description": "Claim an anonymous mixtape, making the authenticated user the owner.",
                "operationId": "claim_mixtape_api_main_mixtape__public_id__claim_post",
                "parameters": [
                    {
                        "name": "public_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Public Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": true,
                                    "title": "Response Claim Mixtape Api Main Mixtape  Public Id  Claim Post"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/main/mixtape/{public_id}": {
            "get": {
                "tags": [
                    "mixtape"
                ],
                "summary": "Get Mixtape",
                "operationId": "get_mixtape_api_main_mixtape__public_id__get",
                "parameters": [
                    {
                        "name": "public_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Public Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MixtapeResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "mixtape"
                ],
                "summary": "Update Mixtape",
                "operationId": "update_mixtape_api_main_mixtape__public_id__put",
                "parameters": [
                    {
                        "name": "public_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Public Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MixtapeRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": true,
                                    "title": "Response Update Mixtape Api Main Mixtape  Public Id  Put"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "MixtapeRequest": {
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 255,
                        "minLength": 1,
                        "title": "Name",
                        "description": "Human-readable name of the mixtape"
                    },
                    "intro_text": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Intro Text",
                        "description": "Optional intro text"
                    },
                    "is_public": {
                        "type": "boolean",
                        "title": "Is Public",
                        "description": "Whether the mixtape is public",
                        "default": false
                    },
                    "tracks": {
                        "items": {
                            "$ref": "#/components/schemas/MixtapeTrackRequest"
                        },
                        "type": "array",
                        "title": "Tracks",
                        "description": "List of tracks in the mixtape"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "tracks"
                ],
                "title": "MixtapeRequest"
            },
            "MixtapeResponse": {
                "properties": {
                    "public_id": {
                        "type": "string",
                        "title": "Public Id"
                    },
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "intro_text": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Intro Text"
                    },
                    "is_public": {
                        "type": "boolean",
                        "title": "Is Public"
                    },
                    "create_time": {
                        "type": "string",
                        "title": "Create Time"
                    },
                    "last_modified_time": {
                        "type": "string",
                        "title": "Last Modified Time"
                    },
                    "stack_auth_user_id": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Stack Auth User Id"
                    },
                    "tracks": {
                        "items": {
                            "$ref": "#/components/schemas/MixtapeTrackResponse"
                        },
                        "type": "array",
                        "title": "Tracks"
                    }
                },
                "type": "object",
                "required": [
                    "public_id",
                    "name",
                    "intro_text",
                    "is_public",
                    "create_time",
                    "last_modified_time",
                    "stack_auth_user_id",
                    "tracks"
                ],
                "title": "MixtapeResponse"
            },
            "MixtapeTrackRequest": {
                "properties": {
                    "track_position": {
                        "type": "integer",
                        "exclusiveMinimum": 0.0,
                        "title": "Track Position",
                        "description": "Unique position of the track within the mixtape (1-based index)"
                    },
                    "track_text": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Track Text",
                        "description": "Optional text to display next to the track"
                    },
                    "spotify_uri": {
                        "type": "string",
                        "maxLength": 255,
                        "minLength": 1,
                        "title": "Spotify Uri",
                        "description": "Spotify URI of the track"
                    }
                },
                "type": "object",
                "required": [
                    "track_position",
                    "spotify_uri"
                ],
                "title": "MixtapeTrackRequest"
            },
            "MixtapeTrackResponse": {
                "properties": {
                    "track_position": {
                        "type": "integer",
                        "title": "Track Position"
                    },
                    "track_text": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Track Text"
                    },
                    "track": {
                        "$ref": "#/components/schemas/TrackDetails"
                    }
                },
                "type": "object",
                "required": [
                    "track_position",
                    "track_text",
                    "track"
                ],
                "title": "MixtapeTrackResponse"
            },
            "TrackAlbum": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "images": {
                        "items": {
                            "$ref": "#/components/schemas/TrackAlbumImage"
                        },
                        "type": "array",
                        "title": "Images"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "images"
                ],
                "title": "TrackAlbum"
            },
            "TrackAlbumImage": {
                "properties": {
                    "url": {
                        "type": "string",
                        "title": "Url"
                    },
                    "width": {
                        "type": "integer",
                        "title": "Width"
                    },
                    "height": {
                        "type": "integer",
                        "title": "Height"
                    }
                },
                "type": "object",
                "required": [
                    "url",
                    "width",
                    "height"
                ],
                "title": "TrackAlbumImage"
            },
            "TrackArtist": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    }
                },
                "type": "object",
                "required": [
                    "name"
                ],
                "title": "TrackArtist"
            },
            "TrackDetails": {
                "properties": {
                    "id": {
                        "type": "string",
                        "title": "Id"
                    },
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "artists": {
                        "items": {
                            "$ref": "#/components/schemas/TrackArtist"
                        },
                        "type": "array",
                        "title": "Artists"
                    },
                    "album": {
                        "$ref": "#/components/schemas/TrackAlbum"
                    },
                    "uri": {
                        "type": "string",
                        "title": "Uri"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "name",
                    "artists",
                    "album",
                    "uri"
                ],
                "title": "TrackDetails"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}